Importing CF Localizations, fixing Addon list naming (#50)
Changelog:

Imported locales from Curseforge Localization.
Removed empty lines and spaces throughout the code.
Adjusted the APR and the APR-Expansions in the Addon list, these should now appear in order again.
Re-added Shadowlands to the Horde list, was accidentally left off.

Expanded `/apr` slash commands:
Added `/apr forcereset` which resets APR's cache of the current character -- Note, this is useful if you have started playing on a character, deleted it, and recreated it. APR saves the current state in your WoW Cache, and ties it to that character. 

Added `/apr showriding` and `/apr hideriding` to show or hide the riding notifications if/when your character gets to a level that can learn a new riding skill.

Added `/apr rollback` to roll back to the previous step (such as back to a waypoint).